A4A: Share the payment method gate across both create site entry points
