+++ client/a8c-for-agencies/sections/purchases/licenses/hooks/use-create-site-from-license.tsx
@@ -1,11 +1,15 @@
 import page from '@automattic/calypso-router';
 import { useCallback, useState } from 'react';
-import { A4A_SITES_LINK_NEEDS_SETUP } from 'calypso/a8c-for-agencies/components/sidebar-menu/lib/constants';
+import {
+	A4A_LICENSES_LINK,
+	A4A_SITES_LINK_NEEDS_SETUP,
+} from 'calypso/a8c-for-agencies/components/sidebar-menu/lib/constants';
 import useFetchPendingSites from 'calypso/a8c-for-agencies/data/sites/use-fetch-pending-sites';
 import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { findPendingSiteIdByLicenseKey, hasProvisioningSite } from '../lib/pending-sites';
 import LicenseSiteConfigurationsModal from '../license-site-configurations-modal';
+import usePaymentMethodGate from './use-payment-method-gate';
 
 type CreateSiteFromLicense = {
 	onCreateSite: () => void;
@@ -18,17 +22,25 @@ type CreateSiteFromLicense = {
  * Lets an unassigned WordPress.com license be turned into a site without
  * leaving the licenses page, using the same configuration modal the Needs
  * setup page opens. Falls back to that page when the license has no pending
- * site to configure — the modal creates a development site when given no ID.
+ * site to configure.
  */
-export default function useCreateSiteFromLicense( licenseKey: string ): CreateSiteFromLicense {
+export default function useCreateSiteFromLicense(
+	licenseKey: string,
+	isClientLicense?: boolean
+): CreateSiteFromLicense {
 	const dispatch = useDispatch();
 	const [ isModalOpen, setIsModalOpen ] = useState( false );
 
 	const { data: pendingSites, isLoading } = useFetchPendingSites();
+	const isBlockedByMissingPaymentMethod = usePaymentMethodGate( isClientLicense );
 
 	const pendingSiteId = findPendingSiteIdByLicenseKey( pendingSites, licenseKey );
 
 	const onCreateSite = useCallback( () => {
+		if ( isBlockedByMissingPaymentMethod( A4A_LICENSES_LINK ) ) {
+			return;
+		}
+
 		if ( ! pendingSiteId ) {
 			dispatch(
 				recordTracksEvent( 'calypso_a4a_licenses_create_site_redirect_needs_setup', {
@@ -45,7 +57,7 @@ export default function useCreateSiteFromLicense( licenseKey: string ): CreateSi
 			} )
 		);
 		setIsModalOpen( true );
-	}, [ dispatch, licenseKey, pendingSiteId ] );
+	}, [ dispatch, isBlockedByMissingPaymentMethod, licenseKey, pendingSiteId ] );
 
 	return {
 		onCreateSite,
