+++ client/a8c-for-agencies/sections/purchases/licenses/lib/pending-sites.ts
@@ -1,7 +1,7 @@
 export type PendingSite = {
 	id: number;
-	features: {
-		wpcom_atomic: {
+	features?: {
+		wpcom_atomic?: {
 			license_key: string;
 			state: string;
 		};
